chore: organized react components into folders
