+++ client/src/services/renderTaskForm.js
@@ -1,5 +1,5 @@
 import React from "react"
-import TaskForm from "../components/TaskForm"
+import TaskForm from "../components/tasks/TaskForm"
 
 const renderTaskForm = (showTaskForm, setShowTaskForm, taskFormProps) => {
   let newTaskFormOutput
